Added Chainlink support
Created the PriceOracle and IPriceOracle interface. This will get the price of different price feeds from Chainlink. It's written in a flexible enough way to add more feeds where necessary. Tournament uses an enum to call the relevant price feed from the PriceOracle. The received price is multiplied to support the 1:100 pegging approah of DAI/WHLE.